// src/shared/api/generated/models/voice-participant-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Voice chat API Gateway
 * Api Gateway for Voice chat application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface VoiceParticipantDto {
    /**
     * ID пользователя
     */
    'userId': string;
    /**
     * ID голосового канала
     */
    'channelId': string;
    /**
     * ID сервера
     */
    'guildId': string;
    /**
     * Момент подключения к каналу
     */
    'joinedAt': string;
}